test: extract shared SharedReference models and cover nested Set parentSet refs
